# swagger.yaml
openapi: 3.0.0
info:
    title: Airbean API
    version: 1.0.0
    description: |
        Airbean backend API for menu, authentication, cart, and orders.
        **Note:** Admin endpoints require a valid JWT token with admin rights.
servers:
    - url: http://localhost:8080/api
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
        Product:
            type: object
            properties:
                prodId:
                    type: string
                title:
                    type: string
                desc:
                    type: string
                price:
                    type: number
                createdAt:
                    type: string
                    format: date-time
                modifiedAt:
                    type: string
                    format: date-time
...
security:
    - bearerAuth: []
paths:
    /menu:
        get:
            summary: Get menu
            description: Returns all items from menu.
            responses:
                '200':
                    description: Menu retrieved
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    menu:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Product'
                '400':
                    description: No menu found

        post:
            summary: Add new product (admin)
            description: Adds a new product to the menu. **Admin only.**
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [title, desc, price]
                            properties:
                                title:
                                    type: string
                                desc:
                                    type: string
                                price:
                                    type: number
            responses:
                '201':
                    description: Product created
                '400':
                    description: Invalid fields
                '401':
                    description: Token missing or invalid
                '403':
                    description: Admin access required
    /menu/{prodId}:
        put:
            summary: Update product (admin)
            description: Updates a product in the menu. **Admin only.**
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: prodId
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                desc:
                                    type: string
                                price:
                                    type: number
            responses:
                '200':
                    description: Product updated
                '400':
                    description: Invalid fields
                '401':
                    description: Token missing or invalid
                '403':
                    description: Admin access required
                '404':
                    description: Product not found

        delete:
            summary: Delete product (admin)
            description: Deletes a product from the menu. **Admin only.**
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: prodId
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Product deleted
                '401':
                    description: Token missing or invalid
                '403':
                    description: Admin access required
                '404':
                    description: Product not found
...